requirements.txt
streamlit
langgraph
chromadb
pandas
sqlalchemy
pytesseract
Pillow
reportlab
python-dotenv
langchain
langchain-community
langchain-google-genai
faiss-cpu
pdf2image
nest_asyncio